+++ raster_feeder/config.py
@@ -24,6 +24,10 @@ STORE_DIR = join(BUILDOUT_DIR, 'var', 'store')
 REDIS_HOST = 'localhost'
 REDIS_DB = 0
 
+# Lizard API credentials
+LIZARD_USERNAME = 'override'
+LIZARD_PASSWORD = 'override'
+
 # Import local settings
 try:
     from .localconfig import *  # NOQA